socket::close() preserves last_error()
As a result of fixing #23 in ddf71320, socket::close() now clears the
last_error property to 0. This means when connector::connect() or
acceptor::open() close themselves after encountering an error, the
error code is lost.
Fixed this by changing socket::close to preserve a pre-existing non-
zero last_error.

Fixes #27